+++ internal/handlers/request_compress.go
