+++ internal/handlers/request_compress_test.go
